refactor transformSaveModelToSchemaV2 and transformSceneToSaveModelV2
