+++ public/app/features/dashboard-scene/scene/layout-tabs/TabsLayoutManager.tsx
@@ -25,6 +25,9 @@ export class TabsLayoutManager extends SceneObjectBase<TabsLayoutManagerState> i
     },
     id: 'tabs-layout',
     createFromLayout: TabsLayoutManager.createFromLayout,
+    getSerializer: () => {
+      throw new Error('Not implemented');
+    },
   };
 
   public readonly descriptor = TabsLayoutManager.descriptor;
