+++ public/app/features/dashboard-scene/serialization/layoutSerializers/layoutSerializerRegistry.ts
@@ -1,0 +1,20 @@
+import { Registry, RegistryItem } from '@grafana/data';
+
+import { LayoutManagerSerializer } from '../../scene/types/DashboardLayoutManager';
+
+import { DefaultGridLayoutManagerSerializer } from './DefaultGridLayoutSerializer';
+import { ResponsiveGridLayoutSerializer } from './ResponsiveGridLayoutSerializer';
+import { RowsLayoutSerializer } from './RowsLayoutSerializer';
+
+interface LayoutSerializerRegistryItem extends RegistryItem {
+  serializer: LayoutManagerSerializer;
+}
+
+export const layoutSerializerRegistry: Registry<LayoutSerializerRegistryItem> =
+  new Registry<LayoutSerializerRegistryItem>(() => {
+    return [
+      { id: 'GridLayout', name: 'Grid Layout', serializer: new DefaultGridLayoutManagerSerializer() },
+      { id: 'ResponsiveGridLayout', name: 'Responsive Grid Layout', serializer: new ResponsiveGridLayoutSerializer() },
+      { id: 'RowsLayout', name: 'Rows Layout', serializer: new RowsLayoutSerializer() },
+    ];
+  });
